Updated JobsModal naming conventions, updated JobsModal to use MaterialUI styling Box component for layout, eliminated the need for scss file
