--- src/styles/Jobs-modal.scss
@@ -1,164 +1,0 @@
-@import "./variables.scss";
-
-.job-modal {
-  &-box {
-    background-color: #F6F6F6;
-  }
-}
-
-#company-logo {
-  width: 100px;
-  height: 100px;
-  margin-right: 15px;
-}
-
-.heading{
-  font-family: 'Montserrat';
-  padding-top: 15px;
-  padding-bottom: 0px;
-  line-height: 1em;
-  margin-top: 0px;
-  margin-bottom: 5px;
-}
-
-.modal-top {
-  display: flex;
-  flex-direction: row;
-  margin: 0;
-}
-
-.modal-top-right {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  align-items: center;
-
-  &-company {
-    width: 31%;
-  }
-
-  &-title {
-    width: 31%;
-  }
-
-  &-status {
-    width: 31%;
-  }
-
-  &-location {
-    width: 31%;
-  }
-
-  &-url {
-    width: 31%;
-  }
-
-  &-salary {
-    width: 31%;
-  }
-}
-
-.logo {
-  border-radius: 0%;
-}
-
-.modal-middle {
-  margin-top: 10px;
-  display: flex;
-  flex-wrap: wrap;
-
-
-  &-details {
-    width: 100%;
-  }
-}
-
-.event {
-  width: 100%;
-  height: 120px;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-
-  &-upcoming {
-    width: 31%;
-  }
-  &-calendar {
-    width: 31%;
-  }
-  &-details {
-    margin-top: 15px;
-    width: 45%;
-  }
-  &-location {
-    width: 31%;
-  }
-
-  &-bottom {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-}
-
-.add-to-calendar {
-  right: 15px;
-  width: 50%;
-  height: 40px;
-  margin-top: 15px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.contact-info {
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.contact {
-  &-name {
-    width: 23%;
-  }
-  &-phone {
-    width: 23%;
-  }
-  &-email {
-    width: 23%;
-  }
-  &-social {
-    width: 23%;
-  }
-}
-
-.links {
-  display: flex;
-  justify-content: space-between;
-
-  &-inputfield {
-    width: 32%;
-  }
-}
-
-.buttons-bottom {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.buttons-right {
-  width: 190px;
-  display: flex;
-  justify-content: space-between;
-}
-
-.buttons-left {
-  width: 200px;
-}
