+++ src/styles/JobsResources.scss
